fs spawn great, using syscall to handle uvpt in user mode, upages to handle the refcount which is a mapping, nor_flash interface, PTE_SHARE for file fd mapping

+++ inc/riscv/types.h
@@ -37,7 +37,8 @@ typedef uint64_t size_t;
 typedef int64_t ssize_t;
 
 // off_t is used for file offsets and lengths.
-typedef int64_t off_t;
+// typedef int64_t off_t;
+typedef int32_t off_t; // in case of compability
 
 // Efficient min and max operations
 #define MIN(_a, _b)						\
